requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
prompt-toolkit>=3.0.0
pyyaml>=6.0
openai>=1.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
